Split optimization_methods into separate files in folder
