+++ src/openqaoa-core/optimizers/optimization_methods/__init__.py
@@ -1,0 +1,8 @@
+from .grad_descent import grad_descent
+from .rmsprop import rmsprop
+from .newton_descent import newton_descent
+from .natural_grad_descent import natural_grad_descent
+from .SPSA import SPSA
+# from .stochastic_grad_descent import stochastic_grad_descent
+from .CANS import CANS
+from .iCANS import iCANS
